separate out template for extension buttons
