+++ app/components/extension-buttons/template.html
@@ -1,0 +1,35 @@
+<div ng-if="$ctrl.isChromium">
+  <a class="btn btn-primary btn-block ph-md-margin-bottom"
+     href="https://chrome.google.com/webstore/detail/paperhive/fihafdlllifbanclcjljledeifcdjbok"
+     onclick="chrome.webstore.install();return false;"
+     >
+    <img
+      alt="Chrome logo"
+      width="30px"
+      src="./static/img/chrome.svg"
+      >
+    Add to Chrome
+  </a>
+  <a href="https://addons.mozilla.org/en-US/firefox/addon/paperhive/">
+    <small>Also available for Mozilla Firefox.</small>
+  </a>
+</div>
+<div ng-if="$ctrl.isFirefox">
+  <a class="btn btn-primary btn-block ph-md-margin-bottom"
+     href="https://addons.mozilla.org/en-US/firefox/addon/paperhive/"
+     >
+    <img
+      alt="Firefox logo"
+      width="30px"
+      src="./static/img/firefox.svg"
+      >
+    Add to Firefox
+  </a>
+  <a href="https://chrome.google.com/webstore/detail/paperhive/fihafdlllifbanclcjljledeifcdjbok">
+    <small>Also available for Google Chrome.</small>
+  </a>
+</div>
+<span ng-if="!$ctrl.isChromium && !$ctrl.isFirefox">
+  Only available for <a href="https://chrome.google.com/webstore/detail/paperhive/fihafdlllifbanclcjljledeifcdjbok">
+  Google Chrome</a> and <a href="https://addons.mozilla.org/en-US/firefox/addon/paperhive/">Mozilla Firefox</a>.
+</span>
