+++ app/components/index.ts
@@ -1,7 +1,7 @@
 import avatar from './avatar';
 import avatarList from './avatar-list';
 import discussionThreadView from './discussion-thread-view/index';
-import extensionButtons from './extension-buttons';
+import extensionButtons from './extension-buttons/index';
 import comment from './comment/index';
 import feedbackButton from './feedback-button';
 import hiveButton from './hive-button';
